refactor(component-version): Moved ComponentVersionIdUtil into a 'deleteme' package.
