+++ blackduck-artifactory-common/src/test/java/com/synopsys/integration/blackduck/artifactory/modules/inspection/notifications/ComponentVersionIdUtilTest.java
@@ -5,6 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import org.junit.jupiter.api.Test;
 
+import com.synopsys.integration.blackduck.artifactory.modules.inspection.notifications.deleteme.ComponentVersionIdUtil;
 import com.synopsys.integration.exception.IntegrationException;
 
 class ComponentVersionIdUtilTest {
